backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0
dj-database-url>=2.3,<3.0
psycopg2-binary>=2.9,<3.0
bleach>=6.1,<7.0
Pillow>=11.0,<12.0
